Added Initial Files with Basic Functionality
Node.java:
Contains all code for a intercetion in a road. This contains the code for the
path finding algorithm

Connection.java:
A class that says which nodes are connected to whom

Main.java:
Used to draw stuff to the screen. Used JavaFX because I don't know of any other graphical libraries so if you want to remove JavaFX I do not mind.

makefile:
Compiles all code and runs it.
Make sure to change JAVAFX_PATH to where the JavaFX lib folder is stored