Added Better image loading and rendering functions
Added a ImageLoader class that will load the image and handle any errors. This will save space in Main.java so that we do not clutter it
ClickableImage is a way to load a image and also give it a rendering object so that you dont have to do that sperately. It also can do an action when clicked on if specifed. The only downside to this type is that if you have an image with transparancy and you click on the transparent pixels, the engine will not register that as a click. To avoid this, use a ImageButton instead. If your image has no transparnecy then use this type instead as it is better than a ImageButton
ImageButton is yet again a easier way to construct a button that has an image.
Changed Main.java to have better image drawing functions that also suport the new ClickableImage type